requirements.txt
google-adk[database,litellm]
yfinance
ollama
psutil
litellm
google-generativeai
python-dotenv